--- src/components/form/Input/ErrorMessage/ErrorMessage.styles.js
@@ -1,15 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { ALERT } from 'constants/colors';
-
-const styles = StyleSheet.create({
-  container: {
-    minHeight: 14,
-    marginTop: 5,
-  },
-  message: {
-    color: ALERT,
-  },
-});
-
-export default styles;
