--- src/components/form/Input/ErrorMessage/index.js
@@ -1,1 +1,0 @@
-export { default } from './ErrorMessage';
