--- src/components/form/Input/Label/index.js
@@ -1,1 +1,0 @@
-export { default } from './Label';
