--- src/components/form/Input/index.js
@@ -1,1 +1,0 @@
-export { default } from './Input';
