--- src/components/form/index.js
@@ -1,4 +1,0 @@
-import ErrorView from './ErrorView';
-import Input from './Input';
-
-export { ErrorView, Input };
