+++ src/components/index.js
@@ -1,4 +1,5 @@
 import BottomSheet from './BottomSheet';
+import BulletItem from './BulletItem';
 import Button from './Button';
 import Card from './Card';
 import Loader from './Loader';
@@ -7,6 +8,4 @@ import SignOut from './SignOut';
 import TabIcon from './TabIcon';
 import Text from './Text';
 
-export * from './form';
-
-export { BottomSheet, Button, Card, Loader, ModalCard, SignOut, TabIcon, Text };
+export { BottomSheet, BulletItem, Button, Card, Loader, ModalCard, SignOut, TabIcon, Text };
