+++ src/httpClient/applyDefaultInterceptors.js
@@ -36,11 +36,7 @@ export default (store, client) => {
         };
         store.dispatch(updateSession(session));
       }
-      response.data = humps.camelizeKeys(data, {
-        process: (key, convert, options) => {
-          return /^[0-9]/g.test(key) ? key : convert(key, options);
-        },
-      });
+      response.data = humps.camelizeKeys(data);
       return response;
     },
     error => {
