+++ src/screens/OneOnOneScreen/OneOnOneInformation/index.js
@@ -1,0 +1,1 @@
+export { default } from './OneOnOneInformation';
