--- src/validations/loginValidations.js
@@ -1,9 +1,0 @@
-export default {
-  email: {
-    presence: true,
-    email: true,
-  },
-  password: {
-    presence: true,
-  },
-};
