--- src/validations/signUpValidations.js
@@ -1,13 +1,0 @@
-export default {
-  email: {
-    presence: true,
-    email: true,
-  },
-  password: {
-    presence: true,
-  },
-  passwordConfirmation: {
-    presence: true,
-    equality: { attribute: 'password' },
-  },
-};
